Creating a new intrinsic "GenISA_getSR0_0 placeholder for sr0.0 the first dword state register, which assumes that sr0.0 is fully Read-Only for performing CSE optimizationn on the calls.

+++ IGC/Compiler/CISACodeGen/EmitVISAPass.hpp
@@ -440,7 +440,7 @@ public:
 
     void emitUnmaskedRegionBoundary(bool start);
     // Debug Built-Ins
-    void emitStateRegID(uint64_t and_imm, uint64_t shr_imm);
+    void emitStateRegID(uint32_t and_imm, uint32_t shr_imm);
     void emitThreadPause(llvm::GenIntrinsicInst* inst);
 
     void MovPhiSources(llvm::BasicBlock* bb);
